# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "skywatch station",
    "description": "The complete contract for a skywatch station: a receive-only SDR aviation\nmonitor that records VHF airband transmissions, transcribes them locally,\nflags the interesting ones, and attaches probable aircraft.\n\nReading the data:\n\n- `/status` \u2014 station health; `/frequencies` \u2014 the listening plan.\n- `/recordings` \u2014 the clip library; `/clips/interesting` \u2014 the good bits;\n  `/digest` \u2014 one day summarised for a \"today\" view.\n- `/runbook` and `/glossary` \u2014 station documents as Markdown.\n\nLive updates come from `WS /stream`. Every message is an envelope\n`{\"type\": <event>, \"payload\": <object>}` with three event types:\n`recording.new` (payload: a /recordings list item), `recording.updated`\n(same shape, sent when a clip's pipeline stage or verdicts change), and\n`status.changed` (payload: the /status shape, sent when station settings\nchange). The stream sends events only \u2014 anything a client sends is ignored.\n\nErrors are RFC 7807 problem details (`application/problem+json`) with a\nstable `code` field to switch on.\n\nThe API is unauthenticated by design: it is meant for a private LAN or\nTailscale network only, and must not be exposed to the public internet.\nTransmissions are received for personal monitoring; audio must not be\nrebroadcast.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/frequencies/{freq_id}/activate": {
      "post": {
        "tags": [
          "frequencies"
        ],
        "summary": "Start recording a frequency",
        "description": "Adds the frequency to the active plan, rewrites the capture configuration, and restarts capture (a restart drops a few seconds of audio \u2014 there is no hot retune). In multichannel mode every active frequency must fit one 2.56 MHz tuner window; when the new set does not fit, the response is a 409 problem detail whose `offenders` lists the frequencies to deactivate and whose `suggested_centerfreq_mhz` is the centre frequency for the largest set that does fit. Switching the station to scan mode is the alternative escape hatch, at the cost of missing concurrent transmissions. Activating an already-active frequency changes nothing.",
        "operationId": "activate_frequency_frequencies__freq_id__activate_post",
        "parameters": [
          {
            "name": "freq_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Freq Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrequencyActionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown frequency.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "The active set would not fit one tuner window.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/frequencies/{freq_id}/deactivate": {
      "post": {
        "tags": [
          "frequencies"
        ],
        "summary": "Stop recording a frequency",
        "description": "Removes the frequency from the active plan, rewrites the capture configuration, and restarts capture. Deactivating the last active frequency stops capture entirely (with a warning in the response) until something is activated again. Deactivating an already-inactive frequency changes nothing.",
        "operationId": "deactivate_frequency_frequencies__freq_id__deactivate_post",
        "parameters": [
          {
            "name": "freq_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Freq Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrequencyActionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown frequency.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/clips/interesting": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "The clips worth hearing",
        "description": "Clips whose latest verdict is interesting, newest first \u2014 the default library view. Equivalent to /recordings?interesting=true, kept as its own route because it is the page most clients open first.",
        "operationId": "list_interesting_clips_clips_interesting_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Clips per page.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Clips per page."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Clips to skip from the newest.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Clips to skip from the newest."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/runbook": {
      "get": {
        "tags": [
          "docs"
        ],
        "summary": "The station runbook",
        "description": "Operating instructions for the station \u2014 start/stop, antenna placement, tuning, troubleshooting \u2014 as raw Markdown for the client to render.",
        "operationId": "get_runbook_runbook_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Runbook not installed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/glossary": {
      "get": {
        "tags": [
          "docs"
        ],
        "summary": "The aviation glossary",
        "description": "Plain-English explanations of the aviation terms that appear in transcripts and classifications, as raw Markdown for the client to render.",
        "operationId": "get_glossary_glossary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Glossary not installed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIErrorCode": {
        "type": "string",
        "enum": [
          "validation_error",
          "not_found",
          "frequency_not_found",
          "recording_not_found",
          "document_not_found",
          "audio_deleted",
          "audio_file_missing",
          "window_conflict",
          "unknown_setting_key",
          "invalid_setting_value",
          "recording_not_classifiable",
          "method_not_allowed",
          "internal_error"
        ],
        "title": "APIErrorCode",
        "description": "Stable machine-readable error codes for problem responses."
      },
...
      "ClassificationCategory": {
        "type": "string",
        "enum": [
          "routine",
          "emergency",
          "urgency",
          "go_around",
          "medical",
          "fuel",
          "guard_activity",
          "unusual",
          "other"
        ],
        "title": "ClassificationCategory",
        "description": "Why a clip is (or is not) interesting."
      },
      "ClassificationResource": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_interesting": {
            "type": "boolean",
            "title": "Is Interesting"
          },
          "category": {
            "$ref": "#/components/schemas/ClassificationCategory"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "description": "Verdict confidence in [0, 1]."
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Plain-language explanation of the verdict."
          },
          "source": {
            "$ref": "#/components/schemas/ClassificationSource"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "LLM model name when the source is llm."
          },
          "prefilter_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prefilter Flags",
            "description": "Which prefilter rules fired (e.g. keyword:mayday, guard_frequency)."
          },
          "status": {
            "$ref": "#/components/schemas/ClassificationStatus",
            "description": "final, or deferred when the daily LLM budget was exhausted and a fuller verdict will be backfilled after the quota resets."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "is_interesting",
          "category",
          "confidence",
          "reason",
          "source",
          "model",
          "prefilter_flags",
          "status",
          "created_at"
        ],
        "title": "ClassificationResource",
        "description": "One routine/interesting verdict. History is append-only; the newest\nverdict is the one that counts."
      },
      "ClassificationSource": {
        "type": "string",
        "enum": [
          "prefilter",
          "llm"
        ],
        "title": "ClassificationSource",
        "description": "Which layer produced a classification verdict."
      },
      "ClassificationStatus": {
        "type": "string",
        "enum": [
          "final",
          "deferred"
        ],
        "title": "ClassificationStatus",
        "description": "Whether a verdict is settled or awaiting an LLM pass after quota reset."
      },
...
      "DocumentResponse": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links",
            "description": "HAL links: self plus related resources and the actions currently available on this resource."
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "markdown": {
            "type": "string",
            "title": "Markdown"
          }
        },
        "type": "object",
        "required": [
          "name",
          "markdown"
        ],
        "title": "DocumentResponse",
        "description": "A station document as raw Markdown for the client to render."
      },
      "FeedbackCounts": {
        "properties": {
          "up": {
            "type": "integer",
            "title": "Up"
          },
          "down": {
            "type": "integer",
            "title": "Down"
          }
        },
        "type": "object",
        "required": [
          "up",
          "down"
        ],
        "title": "FeedbackCounts",
        "description": "Thumbs-up / thumbs-down tallies for a clip."
      },
...
      "FrequencyActionResponse": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links",
            "description": "HAL links: self plus related resources and the actions currently available on this resource."
          },
          "frequency": {
            "$ref": "#/components/schemas/FrequencyResource"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Non-fatal notes about the new plan, e.g. the scan-mode concurrency trade-off or a capture restart that could not complete."
          },
          "capture_restarted": {
            "type": "boolean",
            "title": "Capture Restarted",
            "description": "Whether the capture process was restarted onto the new plan."
          }
        },
        "type": "object",
        "required": [
          "frequency",
          "warnings",
          "capture_restarted"
        ],
        "title": "FrequencyActionResponse",
        "description": "Result of activating or deactivating a frequency."
      },
      "FrequencyCategory": {
        "type": "string",
        "enum": [
          "guard",
          "tower",
          "ground",
          "approach",
          "radar",
          "atis",
          "area_control",
          "airfield"
        ],
        "title": "FrequencyCategory",
        "description": "What kind of aviation service a frequency carries."
      },
...
      "FrequencyMode": {
        "type": "string",
        "enum": [
          "am"
        ],
        "title": "FrequencyMode",
        "description": "Demodulation mode for a monitored frequency."
      },
      "FrequencyRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "mhz": {
            "type": "number",
            "title": "Mhz"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "mhz"
        ],
        "title": "FrequencyRef",
        "description": "Just enough about a frequency to label a clip."
      },
      "FrequencyResource": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links",
            "description": "HAL links: self plus related resources and the actions currently available on this resource."
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "mhz": {
            "type": "number",
            "title": "Mhz"
          },
          "mode": {
            "$ref": "#/components/schemas/FrequencyMode"
          },
          "facility": {
            "type": "string",
            "title": "Facility"
          },
          "category": {
            "$ref": "#/components/schemas/FrequencyCategory"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the station currently records this frequency."
          },
          "tuner_group": {
            "type": "integer",
            "title": "Tuner Group"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified",
            "description": "False until the frequency has been checked against a current official or enthusiast source; unverified rows deserve a prominent caveat in any UI."
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "mhz",
          "mode",
          "facility",
          "category",
          "description",
          "is_active",
          "tuner_group",
          "verified"
        ],
        "title": "FrequencyResource",
        "description": "One entry in the station's frequency plan."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Link": {
        "properties": {
          "href": {
            "type": "string",
            "title": "Href"
          }
        },
        "type": "object",
        "required": [
          "href"
        ],
        "title": "Link",
        "description": "One HAL link: a relative URL on this API."
      },
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "$ref": "#/components/schemas/APIErrorCode"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetail",
        "description": "RFC 7807 problem details; the body of every error response."
      },
...
      "RecordingListResponse": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links",
            "description": "HAL links: self plus related resources and the actions currently available on this resource."
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecordingSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Matching clips across all pages."
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "RecordingListResponse"
      },
      "RecordingStage": {
        "type": "string",
        "enum": [
          "captured",
          "enriching",
          "transcribing",
          "transcribed",
          "classifying",
          "classified",
          "failed_enrich",
          "failed_transcribe",
          "failed_classify"
        ],
        "title": "RecordingStage",
        "description": "Lifecycle position of a captured clip in the processing pipeline.\n\nEnrichment failures are recorded but never block transcription or\nclassification; the failed_* stages mark where a clip needs attention."
      },
      "RecordingSummary": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links",
            "description": "HAL links: self plus related resources and the actions currently available on this resource."
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "frequency": {
            "$ref": "#/components/schemas/FrequencyRef"
          },
          "started_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Started At Utc"
          },
          "ended_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Ended At Utc"
          },
          "duration_s": {
            "type": "number",
            "title": "Duration S"
          },
          "stage": {
            "$ref": "#/components/schemas/RecordingStage"
          },
          "audio_available": {
            "type": "boolean",
            "title": "Audio Available",
            "description": "False once retention has pruned the audio; the row itself is kept."
          },
          "transcript_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript Snippet",
            "description": "Opening of the latest transcript, or null while transcription is pending."
          },
          "classification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClassificationResource"
              },
              {
                "type": "null"
              }
            ],
            "description": "Latest verdict, or null while classification is pending."
          },
          "top_match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AircraftMatchResource"
              },
              {
                "type": "null"
              }
            ],
            "description": "Most plausible aircraft candidate, or null when none was found."
          },
          "feedback": {
            "$ref": "#/components/schemas/FeedbackCounts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "frequency",
          "started_at_utc",
          "ended_at_utc",
          "duration_s",
          "stage",
          "audio_available",
          "transcript_snippet",
          "classification",
          "top_match",
          "feedback"
        ],
        "title": "RecordingSummary",
        "description": "A clip as it appears in lists and cards."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}